push

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: oeddamou <oeddamou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/04 11:27:38 by oeddamou          #+#    #+#             */
-/*   Updated: 2024/12/04 11:51:52 by oeddamou         ###   ########.fr       */
+/*   Updated: 2024/12/07 13:48:29 by oeddamou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ char	*get_next_line(int fd);
 char	*ft_read_line(int fd, char **line);
 int		ft_strjoin(char **buf, char *r);
 char	*ft_substr(char **r);
-int		ft_strchr(char *p);
-int		ft_strcpy(char *d, char *s, int c);
+long	ft_strchr(char *p);
+long	ft_strcpy(char *d, char *s, int c);
 
 #endif
